Make psi interpolation method settable with an option
Make both TORPEXMagneticField and TokamakEquilibrium able to use either
bi-cubic-spline or DCT interpolation, controlled by a new option
psi_interpolation_method (defaulting to spline). Either type can now be
created by the Equilibrium.magneticFunctionsFromGrid() method.

The location where psi is transposed to use DCT has changed (now inside
the DCT_2D.__init__()), which required some changes to the unit tests.

MultiLocationArray was moved to a separate file so it can be imported in
both equilibrium.py and mesh.py.
