+++ hypnotoad/test_suite/test_dct.py
@@ -39,12 +39,12 @@ def test_DCT_2D():
     R_array = numpy.linspace(Rmin, Rmax, nR)
     Z_array = numpy.linspace(Zmin, Zmax, nZ)
 
-    f_array = f(R_array[numpy.newaxis, :], Z_array[:, numpy.newaxis])
+    f_array = f(R_array[:, numpy.newaxis], Z_array[numpy.newaxis, :])
 
     f_dct = DCT_2D(R_array, Z_array, f_array)
 
     # check the input array values are correctly reproduced
-    f_reconstructed = f_dct(R_array[numpy.newaxis, :], Z_array[:, numpy.newaxis])
+    f_reconstructed = f_dct(R_array[:, numpy.newaxis], Z_array[numpy.newaxis, :])
     assert f_reconstructed == tight_approx(f_array)
 
     # check a few random points
@@ -77,24 +77,24 @@ def test_DCT_2D_ddR():
     R_array = numpy.linspace(Rmin, Rmax, nR)
     Z_array = numpy.linspace(Zmin, Zmax, nZ)
 
-    f_array = f(R_array[numpy.newaxis, :], Z_array[:, numpy.newaxis])
-    dfdR_array = dfdR(R_array[numpy.newaxis, :], Z_array[:, numpy.newaxis])
+    f_array = f(R_array[:, numpy.newaxis], Z_array[numpy.newaxis, :])
+    dfdR_array = dfdR(R_array[:, numpy.newaxis], Z_array[numpy.newaxis, :])
 
     f_dct = DCT_2D(R_array, Z_array, f_array)
 
     # check on the input array
-    dfdR_reconstructed = f_dct.ddR(R_array[numpy.newaxis, :], Z_array[:, numpy.newaxis])
+    dfdR_reconstructed = f_dct.ddR(R_array[:, numpy.newaxis], Z_array[numpy.newaxis, :])
     # exclude edge points because gradient reconstruction is poor there
-    assert dfdR_reconstructed[:, nR // 10 : -nR // 10] == pytest.approx(
-        dfdR_array[:, nR // 10 : -nR // 10], abs=1.0e-2
+    assert dfdR_reconstructed[nR // 10 : -nR // 10, :] == pytest.approx(
+        dfdR_array[nR // 10 : -nR // 10, :], abs=1.0e-2
     )
 
     # check a few random points
     R = numpy.array([0.2784230, 0.357892, 0.578237, 0.732580, 1.0326794])[
-        numpy.newaxis, :
+        :, numpy.newaxis
     ]
     Z = numpy.array([-0.232123, -0.178594, -0.053789, 0.172530, 0.375072])[
-        :, numpy.newaxis
+        numpy.newaxis, :
     ]
     # can't use tight tolerance because interpolation does not reproduce exactly the
     # values away from the grid points
@@ -119,24 +119,24 @@ def test_DCT_2D_ddZ():
     R_array = numpy.linspace(Rmin, Rmax, nR)
     Z_array = numpy.linspace(Zmin, Zmax, nZ)
 
-    f_array = f(R_array[numpy.newaxis, :], Z_array[:, numpy.newaxis])
-    dfdZ_array = dfdZ(R_array[numpy.newaxis, :], Z_array[:, numpy.newaxis])
+    f_array = f(R_array[:, numpy.newaxis], Z_array[numpy.newaxis, :])
+    dfdZ_array = dfdZ(R_array[:, numpy.newaxis], Z_array[numpy.newaxis, :])
 
     f_dct = DCT_2D(R_array, Z_array, f_array)
 
     # check on the input array
-    dfdZ_reconstructed = f_dct.ddZ(R_array[numpy.newaxis, :], Z_array[:, numpy.newaxis])
+    dfdZ_reconstructed = f_dct.ddZ(R_array[:, numpy.newaxis], Z_array[numpy.newaxis, :])
     # exclude edge points because gradient reconstruction is poor there
-    assert dfdZ_reconstructed[nZ // 10 : -nZ // 10] == pytest.approx(
-        dfdZ_array[nZ // 10 : -nZ // 10], abs=1.0e-2
+    assert dfdZ_reconstructed[:, nZ // 10 : -nZ // 10] == pytest.approx(
+        dfdZ_array[:, nZ // 10 : -nZ // 10], abs=1.0e-2
     )
 
     # check a few random points
     R = numpy.array([0.2784230, 0.357892, 0.578237, 0.732580, 1.1326794])[
-        numpy.newaxis, :
+        :, numpy.newaxis
     ]
     Z = numpy.array([-0.222123, -0.178594, -0.053789, 0.172530, 0.275072])[
-        :, numpy.newaxis
+        numpy.newaxis, :
     ]
     # can't use tight tolerance because interpolation does not reproduce exactly the
     # values away from the grid points
